Implemented Slider on Home and Bookkeeping pages
Added slider to home and bookkeeping pages, also added custom captions
and most of the images are very lightly colour/lit and therefore i
thought the phrases needed some differentiation. The text is now
actually text which will be picked up by google in crawling (rather than
embedded in an image where it is not noticed.). Took the extra time to
fix the typography on the single slides at present.